[Refactor] Add pageInfo field in retrieveAllContest api response

+++ src/main/java/dayum/dayumserver/domain/contents/ContentsRepository.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface ContentsRepository {
 
-  List<Contents> fetchNextPage(long previousId, long size);
+  List<Contents> fetchNextPage(long previousId, int size);
 }
